fix mypy, lint, flake8
