+++ checkov/common/parsers/multiline_parser.py
@@ -5,18 +5,19 @@ from abc import ABC
 from abc import abstractmethod
 
 if TYPE_CHECKING:
-	from detect_secrets.util.code_snippet import CodeSnippet
+    from detect_secrets.util.code_snippet import CodeSnippet
 
 
 class BaseMultiLineParser(ABC):
-	def __init__(self):
-		pass
+    @abstractmethod
+    def __init__(self) -> None:
+        pass
 
-	@abstractmethod
-	def get_lines_from_same_object(
-			self,
-			search_range: range,
-			context: CodeSnippet | None,
-			raw_context: CodeSnippet | None
-	):
-		pass
+    @abstractmethod
+    def get_lines_from_same_object(
+        self,
+        search_range: range,
+        context: CodeSnippet | None,
+        raw_context: CodeSnippet | None
+    ) -> set[str]:
+        pass
